fix[test]: flaky test
It seems the tests are flaky if the notifier is constantly stopped and
started. We alredy set disable teh notifier stop in the jenkins tests.
Let's do this here also.

I guess the tests should work in both cases, but I currently don't have
a better idea.

Issue univention/ucsschool#1372