feat(u-c-umc-exam): update permissions of PDFprinter queue
When exam users get removed and recreated, a new uidNumber is assigned
to the exam users. The user specific printer queue of PDFprinter is not
recreated and keeps the old directory owner that no longer exists.
To prevent printing problems the exam module now also corrects the
permissions of /var/spool/cups-pdf/USERNAME/.

Issue univention/ucsschool#1408